requirements.txt

celery==5.2.7
channels==3.0.5
channels-redis==3.4.0
Django
django-celery-results==2.3.1
django-crontab==0.7.1
django-debug-toolbar==3.4.0
django-dotenv==1.4.2
django-environ==0.8.1
django-redis==5.2.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
python-dotenv==0.20.0
PyYAML==5.3.1
redis==4.3.4


numpy
pandas
